ConfigValueProviderTransformer: added ability to tolerate invalid config reference specs.
Sometimes valid loaded configurations contain substrings in a form of
`{foo}`, for example `x: "hello, {name}!" - in such cases `{name}` was
mistaken for an old-style config value reference specification which has
a syntax of `{provider[:name]://<config value key>}` and exception was
thrown because of this. This format is obsolete and will be removed in
the future.

This PR adds exception in such cases and makes behavious configurable in
the code.