[aws2-sts]: forcibly include aws2 sts webidentity into native-image
AWS EKS uses web-identity tokens to implement fine-grained ACLs and
support for that feature is part of aws2-sts dependency, but because
it's being loaded using reflection it doesn't get included in the
produced binary by default. This patch forces inclusion into binary by
marking it open for reflective access.